Changed compiler
Forgot how java packaging worked and compiling classes from the src/ directory made the location of the classes be dist/src/[class file], which was not great. Changed now to simply compile .java files from the root directory for the assignment and place the class files into a classes/ directory.