An attempt at implementation without hospital rankings
Currently "works". The stability of the matches are to be debated. Basically how it works is first-come first serve. Hospitals are filled based on the first choice of the resident. If their first choice is full, they get their second choice. If that is full, they get their second, and so on.